+++ src/Components/PaymentMethod/CouponCodeField.js
@@ -1,22 +1,33 @@
 import React, { useContext } from "react";
 import { useTranslation } from "react-i18next";
 import { store } from "./PaymentMethodContainer";
-import { UPDATE_COUPON_CODE } from "../../utils/action-types";
+import {
+  SET_COUPON_ERROR,
+  UPDATE_COUPON_CODE
+} from "../../utils/action-types";
 import { Input } from "../../SubComponents/Input";
 
 export const CouponCodeField = (props) => {
   const { t } = useTranslation("checkoutForm");
   const {
     dispatch,
-    state: { couponCode }
+    state: { couponCode, couponError }
   } = useContext(store);
 
-  const onCouponCodeChange = (e) =>
+  const onCouponCodeChange = (e) => {
     dispatch({ type: UPDATE_COUPON_CODE, payload: e.target.value });
+
+    if (couponError) {
+      dispatch({ type: SET_COUPON_ERROR, payload: "" });
+    }
+  };
 
   return (
     <Input
-      hideErrors={true}
+      className="plc-h-12"
+      errorClassName="plc-h-12 sm:plc-h-8"
+      hideErrors={!couponError}
+      error={couponError}
       aria-label={t("labels.code")}
       value={couponCode}
       onChange={onCouponCodeChange}