+++ src/Components/PaymentMethod/PaymentMethodContainer.js
@@ -13,7 +13,6 @@ import useReducerWithSideEffects, {
 import {
   DISABLE_SUBMIT,
   LOADING,
-  CREATE_PAYMENT,
   SET_UPDATED_PRICE,
   SUBMIT_PAYMENT,
   HANDLE_PAYPAL_SUBSCRIPTION,
@@ -21,6 +20,7 @@ import {
   APPLY_COUPON_CODE,
   SET_PERCENT_OFF,
   SET_COUPON,
+  SET_COUPON_ERROR,
   UPDATE_COUPON_CODE,
   SHOW_COUPON_FIELD,
   SET_CAN_MAKE_PAYMENT,
@@ -48,6 +48,7 @@ import {
  * @property {boolean} isLoading
  * @property {boolean} disableCouponButton
  * @property {string} couponCode
+ * @property {string} couponError
  * @property {boolean} enableCouponField
  * @property {string} percentOff
  * @property {unknown} canMakePayment
@@ -64,6 +65,7 @@ const initialState = {
   isLoading: false,
   disableCouponButton: false,
   couponCode: "",
+  couponError: "",
   enableCouponField: false,
   percentOff: "",
   canMakePayment: false,
@@ -203,14 +205,11 @@ const PaymentMethodContainerWithoutStripe = ({
 
           if (err) {
             dispatch({ type: SET_PERCENT_OFF, payload: "" });
-
             onFailure(err);
+
             return dispatch({
-              type: SHOW_ALERT,
-              payload: {
-                type: "error",
-                content: getErrorMessages(err)
-              }
+              type: SET_COUPON_ERROR,
+              payload: getErrorMessages(err)
             });
           }
 
@@ -560,6 +559,9 @@ const PaymentMethodContainerWithoutStripe = ({
 
         case SET_COUPON:
           return Update({ ...state, coupon: action.payload });
+
+        case SET_COUPON_ERROR:
+          return Update({ ...state, couponError: action.payload });
 
         case UPDATE_COUPON_CODE:
           return UpdateWithSideEffect(