+++ changelog.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog][keep a changelog] and this project adh
 
 ### Changed
 
+- Place coupon related errors under the coupon field.
 - Apply coupon button appearence & Added gap between field and button.
 
 ## [Released]